requirements.txt
Django==3.2.6
django-filter==2.4.0
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
drf-yasg==1.20.0
flake8==3.9.2
model-mommy==2.0.0
psycopg2-binary==2.8.6
python-decouple==3.4